requirements.txt
absl-py==0.11.0
antlr4-python3-runtime==4.8
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
efficientnet-pytorch==0.6.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
hydra-core==1.0.4
idna==2.10
importlib-metadata==3.3.0
importlib-resources==4.0.0
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
munch==2.5.0
numpy==1.19.4
oauthlib==3.1.0
omegaconf==2.0.5
Pillow==8.0.1
pretrainedmodels==0.7.4
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
segmentation-models-pytorch==0.1.3
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
timm==0.3.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
zipp==3.4.0